+++ persistent-template/test/TemplateTestImports.hs
@@ -1,10 +1,16 @@
 {-# LANGUAGE TemplateHaskell #-}
-module TemplateTestImports where
+
+module TemplateTestImports
+    ( module TemplateTestImports
+    , module X
+    ) where
 
 import Data.Aeson.TH
 import Test.QuickCheck
 
-import Database.Persist.TH
+import Test.Hspec as X
+import Database.Persist.Sql as X
+import Database.Persist.TH as X
 
 data Foo = Bar | Baz
     deriving (Show, Eq)